+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased Changes
 * Added `NonMaxI128` and `NonMaxU128` support
+* Implemented `std::convert::[Try]From` for `NonMax*` & primitive types to match `NonZero*`
 
 ## 0.4.0 (2020-09-27)
 * Raised MSRV to 1.46.0 to make more methods `const`.